_customed removed; integrate network into networks file
