+++ examples/networks/actor.py
@@ -97,3 +97,30 @@ class CategoricalActor(nn.Module):
         greedy = torch.argmax(prob, dim=-1).unsqueeze(-1)  # 1d tensor
 
         return sample_action, prob, logprob, greedy
+
+class openai_actor(nn.Module):
+    def __init__(self, num_inputs, action_size):
+        super(openai_actor, self).__init__()
+        self.tanh= nn.Tanh()
+        self.LReLU = nn.LeakyReLU(0.01)
+        self.linear_a1 = nn.Linear(num_inputs, 128)
+        self.linear_a2 = nn.Linear(128, 64)
+        self.linear_a = nn.Linear(64, action_size)
+        self.reset_parameters()
+        self.train()
+    
+    def reset_parameters(self):
+        gain = nn.init.calculate_gain('leaky_relu')
+        gain_tanh = nn.init.calculate_gain('tanh')
+        nn.init.xavier_uniform_(self.linear_a1.weight, gain=nn.init.calculate_gain('leaky_relu'))
+        nn.init.xavier_uniform_(self.linear_a2.weight, gain=nn.init.calculate_gain('leaky_relu'))
+        nn.init.xavier_uniform_(self.linear_a.weight, gain=nn.init.calculate_gain('leaky_relu'))
+    
+    def forward(self, input, original_out=False):
+        x = self.LReLU(self.linear_a1(input))
+        x = self.LReLU(self.linear_a2(x))
+        model_out = self.linear_a(x)
+        u = torch.rand_like(model_out)
+        policy = F.softmax(model_out - torch.log(-torch.log(u)), dim=-1)
+        if original_out == True:   return model_out, policy
+        return policy
